chore: domain split + route-utils cleanup (Phases 1–2)
Split packages/domain/src/index.ts (739 lines) into focused sub-modules:
statuses.ts, entities.ts, pricing.ts, visits.ts — all re-exported from
index.ts for full backward compatibility.

Extracted getPathId() utility to lib/route-utils.ts and applied it across
all 13 dynamic API route files, replacing 24 hand-rolled pathname.split()
calls. Deleted stale vim swap file.
