+++ apps/web/app/api/v1/membership-pricing/[id]/route.ts
@@ -3,6 +3,7 @@ import { z } from "zod";
 import { withRole } from "@/lib/auth/middleware";
 import type { AuthSession } from "@/lib/auth/middleware";
 import { queryOne, getPool } from "@/lib/db";
+import { getPathId } from "@/lib/route-utils";
 
 export const dynamic = "force-dynamic";
 
@@ -14,7 +15,7 @@ const patchBody = z.object({
 });
 
 export const PATCH = withRole(["owner", "admin"], async (request: NextRequest, session: AuthSession) => {
-  const id = request.nextUrl.pathname.split("/").at(-1) ?? "";
+  const id = getPathId(request.nextUrl.pathname);
 
   const existing = await queryOne<{ id: string; tier: string; is_published: boolean } & Record<string, unknown>>(
     `SELECT id, tier, is_published FROM membership_pricing_structures WHERE id = $1 AND account_id = $2`,
@@ -82,7 +83,7 @@ export const PATCH = withRole(["owner", "admin"], async (request: NextRequest, s
 });
 
 export const DELETE = withRole(["owner", "admin"], async (request: NextRequest, session: AuthSession) => {
-  const id = request.nextUrl.pathname.split("/").at(-1) ?? "";
+  const id = getPathId(request.nextUrl.pathname);
 
   const existing = await queryOne<{ is_published: boolean } & Record<string, unknown>>(
     `SELECT is_published FROM membership_pricing_structures WHERE id = $1 AND account_id = $2`,
